requirements.txt
numpy>=1.25
scipy>=1.1
PyYAML>=5.1
